Cleanup for Merge to Master
Cleaned up working files for merge with master branch
