+++ rds-copy-snapshots-lambda_monthly.py
@@ -1,39 +1,34 @@
 from __future__ import print_function
 from dateutil import parser, relativedelta
 from boto3 import client
-import datetime
 
 # List of database identifiers, or "all" for all databases
 # eg. ["db-name1", "db-name2"] or ["all"]
-INSTANCES = ["db-name-01"]
+INSTANCES = ["sql-unanet-01"]
 
 # The number of months to keep ONE snapshot per month
 MONTHS = 14
 
 # three character prefix for snapshot naming : "mo-" = monthly "wk-" = weekly
-str_snap_name_prefx="mo-"
+SNAP_NAME_PREFX="mo-"
 
 # AWS region in which the db instances exist
 REGION = "us-east-1"
 
-# ensure all snaps return a SnapshotCreateTime - Snapshots in progress will not have a timestamp until completed
-def byTimestamp(snap):  
-    if 'SnapshotCreateTime' in snap:
-        return datetime.datetime.isoformat(snap['SnapshotCreateTime'])
-    else:
-        return datetime.datetime.isoformat(datetime.datetime.now())
 
 
-def copy_snapshots(rds, snaps):
+def copy_snapshots(rds, snaps, dbarn):
     newest = snaps[-1]
+    response = rds.list_tags_for_resource(ResourceName=dbarn)
+    rdstags = response['TagList']
+    rdstags.append({ 'Key': 'backup_type', 'Value': 'monthly' })
     rds.copy_db_snapshot(
         SourceDBSnapshotIdentifier=newest['DBSnapshotIdentifier'],
-        TargetDBSnapshotIdentifier=str_snap_name_prefx + newest['DBSnapshotIdentifier'][4:],
-        CopyTags=True
-    )
+        TargetDBSnapshotIdentifier=SNAP_NAME_PREFX + newest['DBSnapshotIdentifier'][4:],
+        Tags= rdstags)
     print("Snapshot {} copied to {}".format(
           newest['DBSnapshotIdentifier'],
-          str_snap_name_prefx + newest['DBSnapshotIdentifier'][4:])
+          SNAP_NAME_PREFX + newest['DBSnapshotIdentifier'][4:])
           )
 
 
@@ -46,21 +41,16 @@ def purge_snapshots(rds, id, snaps, counts):
     print("---- RESULTS FOR {} ({} snapshots) ----".format(id, len(snaps)))
 
     for snap in snaps:
-#        snap_date = snap['SnapshotCreateTime']
-        if 'SnapshotCreateTime' in snap:
-            snap_date = snap['SnapshotCreateTime']
-        else:
-            snap_date = NOW
-            
+        snap_date = snap['SnapshotCreateTime']   
         snap_age = NOW - snap_date
         # Monthly
         type_str = "month"
-        start_date_str = snap_date.strftime("%Y-%m")
+        start_date_str = snap_date.strftime("%Y-%m-%d")
         if (start_date_str != prev_start_date and
                 snap_date > DELETE_BEFORE_DATE):
             # Keep it
             prev_start_date = start_date_str
-            print("Keeping {}: {}, {} unix epoc days old - {} of {}".format(
+            print("Keeping {}: {}, {} days old - {} of {}".format(
                   snap['DBSnapshotIdentifier'], snap_date, snap_age.days,
                   type_str, start_date_str)
                   )
@@ -88,6 +78,19 @@ def purge_snapshots(rds, id, snaps, counts):
     counts[id] = [delete_count, keep_count]
 
 
+def get_snaps_filtered(rds, instance, snap_type):
+    str_status_type="avail"
+    if len(INSTANCES) == INSTANCES.count("all"):
+        snapshots = rds.describe_db_snapshots(
+                    SnapshotType=snap_type)['DBSnapshots']
+    else:
+        snapshots = rds.describe_db_snapshots(
+                    SnapshotType=snap_type,
+                    DBInstanceIdentifier=instance)['DBSnapshots']
+    snapshots=filter(lambda x: x['Status'].startswith(str_status_type), snapshots)  #filter snaps based on status=available - returning only snaps that not creating or deleting
+    snapshots=filter(lambda x: x['DBSnapshotIdentifier'].startswith(SNAP_NAME_PREFX), snapshots)  #filter the snapshots based on the the first letters of the DBSnapshotIdentifier
+    return sorted(snapshots, key=lambda x: x['SnapshotCreateTime'])
+    
 def get_snaps(rds, instance, snap_type):
     if len(INSTANCES) == INSTANCES.count("all"):
         snapshots = rds.describe_db_snapshots(
@@ -96,8 +99,8 @@ def get_snaps(rds, instance, snap_type):
         snapshots = rds.describe_db_snapshots(
                     SnapshotType=snap_type,
                     DBInstanceIdentifier=instance)['DBSnapshots']
-#    return sorted(snapshots, key=lambda x:x['SnapshotCreateTime'])
-    return sorted(snapshots, key=byTimestamp, reverse=False)
+    snapshots=filter(lambda x: x['Status'].startswith('avail'), snapshots)  #filter snaps based on status=available - returning only snaps that not creating or deleting
+    return sorted(snapshots, key=lambda x: x['SnapshotCreateTime'])
 
 
 def print_summary(counts):
@@ -128,16 +131,22 @@ def main(event, context):
             instance_counts = {}
             snapshots_auto = get_snaps(rds, instance, 'automated')
             if snapshots_auto:
-                copy_snapshots(rds, snapshots_auto)
+                print("Processing Snapshots for instance: {} ".format(instance))
+                response = rds.describe_db_instances(DBInstanceIdentifier=instance)
+                dbins = response['DBInstances']
+                dbin = dbins[0]
+                dbarn = dbin['DBInstanceArn']
+                print("The instance arn is:  {} ".format(dbarn))
+                copy_snapshots(rds, snapshots_auto, dbarn)
             else:
                 print("No auto snapshots found for instance: {}".format(
                       instance)
                       )
-            snapshots_manual = get_snaps(rds, instance, 'manual')
+            snapshots_manual = get_snaps_filtered(rds, instance, 'manual')
             if snapshots_manual:
                 print("\nNumber of Monthly Snaps to retain = ",MONTHS," \n" )
-                print("The variable NOW is set to ",NOW," \n")
-                print("Snapshots will be deleted prior to ",DELETE_BEFORE_DATE," \n" ) 
+                print("Script start time is: ",NOW," \n")
+                print("Snapshots will be deleted prior to: ",DELETE_BEFORE_DATE," \n" ) 
                 purge_snapshots(rds, instance,
                                 snapshots_manual, instance_counts)
                 print_summary(instance_counts)
